v3.0 Psr4 + support for prooph/common >=3.x and prooph/event-store >=5.x
