+++ src/EventStoreIntegration/AggregateRootDecorator.php
