+++ src/EventStoreIntegration/AggregateTranslator.php
