+++ tests/EventStoreIntegration/AggregateTranslatorTest.php
